Ported to App Engine
Made necessary changes to run on Google App Engine
