+++ Flask/steamapp/app.yaml
@@ -1,0 +1,1 @@
+runtime: python37
