+++ Flask/steamapp/requirements.txt
@@ -1,0 +1,3 @@
+Flask
+numpy
+scikit-learn
