+++ Flask/steamapp/templates/emptyReview.html
@@ -1,4 +1,9 @@
 <html>
+	<head>
+		<style>
+			html {zoom:200%;}
+		</style>
+	</head>
 	<body style="background: linear-gradient(110deg, #3d7078 60%, #181e26 60%)">
 		<h3 style="color: #d6b97e">You must enter a review before our robots can assess your feelings!</h3>
 		<form action="http://127.0.0.1:5000">
